refactor: resolve SonarQube blocker by eliminating code duplication in models
- Create BaseNamedModelMixin for common fields (id, name, category) in named models
- Create BaseJunctionModelMixin for common fields (id, professional_id) in junction tables
- Refactor therapeutic_approach.py, specialty.py, modality.py to use BaseNamedModelMixin
- Refactor professional_specialty.py, professional_therapeutic_approach.py to use BaseJunctionModelMixin
- Eliminate duplicate code patterns that were causing SonarQube 'Similar lines in 2 files' blocker
- Maintain 100% backward compatibility and functionality
- All tests passing, linters clean (pylint 9.93/10, flake8 0 violations, black formatted)

Resolves: SonarQube code duplication blocker
