+++ backend/app/models/professional_specialty.py
@@ -2,23 +2,19 @@
 Professional Specialty model for the Miamente platform.
 """
 
-import uuid
-
 from sqlalchemy import Column, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 
 from app.core.database import Base
-from app.models.mixins import TimestampMixin
+from app.models.mixins import BaseJunctionModelMixin, TimestampMixin
 
 
-class ProfessionalSpecialty(Base, TimestampMixin):
+class ProfessionalSpecialty(Base, BaseJunctionModelMixin, TimestampMixin):
     """Professional–Specialty link: many-to-many relationship between professionals and specialties."""
 
     __tablename__ = "professional_specialties"
 
-    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    professional_id = Column(UUID(as_uuid=True), ForeignKey("professionals.id"), nullable=False)
     specialty_id = Column(UUID(as_uuid=True), ForeignKey("specialties.id"), nullable=False)
 
     # Relationships
