+++ backend/app/models/professional_therapeutic_approach.py
@@ -5,25 +5,21 @@ This model manages the many-to-many relationship between professionals and their
 therapeutic approaches, allowing professionals to specialize in multiple approaches.
 """
 
-import uuid
-
 from sqlalchemy import Column, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 
 from app.core.database import Base
-from app.models.mixins import TimestampMixin
+from app.models.mixins import BaseJunctionModelMixin, TimestampMixin
 
 
-class ProfessionalTherapeuticApproach(Base, TimestampMixin):
+class ProfessionalTherapeuticApproach(Base, BaseJunctionModelMixin, TimestampMixin):
     """Professional Therapeutic Approach model - Many-to-many relationship between profe
     ssionals and
     therapeutic approaches."""
 
     __tablename__ = "professional_therapeutic_approaches"
 
-    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    professional_id = Column(UUID(as_uuid=True), ForeignKey("professionals.id"), nullable=False)
     therapeutic_approach_id = Column(UUID(as_uuid=True), ForeignKey("therapeutic_approaches.id"), nullable=False)
 
     # Relationships
